FIXED: SMSG_ITEM_QUERY_SINGLE_RESPONSE - last field is HolidayId as of 3.1.3 DELETED: Obsolete and useless columns from items table FIXED: PVP Title The Insane http://arcemu.org/forums/index.php?showtopic=17231 FIXED: Vellums, for more info read this: http://www.wowwiki.com/Vellum FIXED: Client stacking overflow http://arcemu.org/forums/index.php?showtopic=17207 FIXED: Query, when someone delete character http://arcemu.org/forums/index.php?showtopic=17207 FIXED: Soul shard creation: http://arcemu.org/forums/index.php?showtopic=17068 FIXED: Arena Player Counting http://arcemu.org/forums/index.php?showtopic=17125 FIXED: Compile warnings FIXED: Gouge FIXED: Improved Devotion Aura FIXED: Glyph of Seal of Light FIXED: Seal of Light (heal amount) FIXED: Enchant Weapon - Deathfrost FIXED: Sigil of the Unfaltering Knight FIXED: Deadly Gladiator's Death Knight Relic FIXED: Deadly Gladiator's Idol of Resolve FIXED: Deadly Gladiator's Libram of Fortitude FIXED: Deadly Totem of Indomitability FIXED: Sigil of Haunted Dreams FIXED: Vestige of Haldor FIXED: Forge Ember FIXED: Mirror of Truth FIXED: Majestic Dragon Figurine FIXED: Flow of Knowledge FIXED: Embrace of the Spider FIXED: Anvil of Titans FIXED: Soul of the Dead FIXED: Illustration of the Dragon Soul FIXED: Grim Toll FIXED: Fury of the Five Flights FIXED: Bandit's Insignia FIXED: Meteorite Whetstone FIXED: Sonic Booster FIXED: Totem of the Third Wind FIXED: Moonkin Starfire Bonus
Some of those spells are NTY, report any bugs on forums.

+++ src/arcemu-world/LootMgr.cpp
@@ -871,7 +871,7 @@ void LootRoll::Finalize()
 		add->SetCount(add->GetUInt32Value(ITEM_FIELD_STACK_COUNT) + amt);
 		add->m_isDirty = true;
 		sQuestMgr.OnPlayerItemPickup(_player,add);
-		_player->GetSession()->SendItemPushResult(add, false, true, true, false, _player->GetItemInterface()->GetBagSlotByGuid(add->GetGUID()), 0xFFFFFFFF, 1);
+		_player->GetSession()->SendItemPushResult(add, false, true, true, false, (uint8)_player->GetItemInterface()->GetBagSlotByGuid(add->GetGUID()), 0xFFFFFFFF, 1);
 		#ifdef ENABLE_ACHIEVEMENTS
 		_player->GetAchievementMgr().UpdateAchievementCriteria(ACHIEVEMENT_CRITERIA_TYPE_LOOT_ITEM, add->GetEntry(), 1, 0);
 #endif
